+++ bc/recruitment/models.py
@@ -36,7 +36,7 @@ class JobSubcategory(models.Model):
     This corresponds to Job Group in the TalentLink import API
     """
 
-    title = models.CharField(max_length=128)
+    title = models.CharField(max_length=128, unique=True)
 
     def get_categories_list(self):
         if self.categories:
@@ -195,7 +195,7 @@ class TalentLinkJob(models.Model):
     searchable_salary = models.CharField(
         max_length=255, help_text="Salary group for filtering"
     )
-    searchable_location = models.CharField(max_length=255)
+    location = models.CharField(max_length=255)
     location_postcode = models.CharField(max_length=8, blank=True)
     location_lat = models.DecimalField(
         max_digits=9, decimal_places=6, null=True, blank=True