+++ bc/recruitment_api/management/commands/import_jobs.py
@@ -79,7 +79,7 @@ class Command(BaseCommand):
                             )
                         except Exception as e:
                             msg = (
-                                f"Error occurred while processing job {ad['id']}:\n"
+                                f"Error occurred while processing job {ad['id']} {ad['jobNumber']}:\n"
                                 + str(e)
                             )
                             errors.append(msg)
@@ -94,7 +94,7 @@ class Command(BaseCommand):
                                 doc_imported += import_attachments_for_job(job, client)
                             except Exception as e:
                                 msg = (
-                                    f"Error occurred while importing attachments for job {ad['id']}:\n"
+                                    f"Error occurred while importing attachments for job {ad['id']} {ad['jobNumber']}:\n"
                                     + str(e)
                                 )
                                 errors.append(msg)
@@ -133,7 +133,7 @@ class Command(BaseCommand):
                                             break
                             except Exception as e:
                                 msg = (
-                                    f"Error occurred while importing logo image for job {ad['id']}:\n"
+                                    f"Error occurred while importing logo image for job {ad['id']} {ad['jobNumber']}:\n"
                                     + str(e)
                                 )
                                 errors.append(msg)
@@ -166,5 +166,5 @@ class Command(BaseCommand):
             self.stdout.write(f"{image_imported} new images imported")
             self.stdout.write(f"{num_deleted} jobs deleted")
             self.stdout.write(self.style.ERROR(f"{len(errors)} errors"))
-            for error in errors:
+            for msg in errors:
                 self.stdout.write(self.style.ERROR(msg))